requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
langchain-groq
python-dotenv
pydantic<2.0.0
markdown
pymupdf
